Lot of work on #101
Started to work on lining up the actions with the content at larger resolutions. This is even harder to support than I expected (at least for how I'm approaching it).

I've don't have this working for preview pane or multiple inboxes yet and I bet there are some cases where it breaks.