Fix for squished inlined images
Just uploaded a fix for the squished LinkedIn images for non-preview pane views as v1.5.3.

Why they were squished: I was applying a max-width of 100% to keep really large inlined images from forcing the whole screen to scroll horizontally. For some reason, that max-width squished profile photos on LinkedIn invites.

Why it isn't fixed for preview pane: I'm now calculating the max-width of the content and setting the max-width to a pixel value but Preview Pane has a dynamic content width that I need to do more to detect and set.