Fix for #250
A fix for when Preview Pane is enabled but toggled off (set to "No Split"). This restores the max-width on the view and adds a warning that people that don't want to use Preview Pane should disable it in Gmail Settings as even with this fix, it still is a sub-par experience.